Reorganise alert functions into a module
