+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2018-11-08
+
+* [BREAKING] Moved `config.alert` to `config.alert.email`
+
 # 2017-07-24
 
 * [BREAKING] Dashboard server specification no longer uses the URL hash. It now uses a `server` URL query parameter `http://rcanary.example.com?server=ws://localhost:8099`
