+++ librcanary/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "librcanary"
-version = "0.2.0"
+version = "0.4.0"
 authors = ["Ng Guoyou <ng.guoyou@gmail.com>"]
 
 [lib]
